Split HostActivity to HostActivity, GameStateType, HostActivityEvents and HostActivityConstants.
