+++ sources/Colonization/Activities/HostActivity/GameStateType.hpp
@@ -1,0 +1,11 @@
+#pragma once
+namespace Colonization
+{
+enum GameStateType
+{
+    GAME_STATE_UNITIALIZED = 0,
+    GAME_STATE_WAITING_FOR_START = 1,
+    GAME_STATE_PLAYING = 2,
+    GAME_STATE_FINISHED = 3
+};
+}
