+++ sources/Colonization/Activities/HostActivity/HostActivityConstants.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include <Urho3D/Math/StringHash.h>
+
+namespace Colonization
+{
+const Urho3D::StringHash VAR_REPLICATED_NODES_COUNT ("ReplicatedNodesCount");
+const Urho3D::StringHash VAR_MAP_FOLDER ("MapFolder");
+const Urho3D::StringHash VAR_MAP_INFO_PATH ("MapInfoPath");
+}
