+++ sources/Colonization/AngelScriptBinders/Activities/HostActivity/BindGameStateType.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindGameStateType(Urho3D::Script *script);
+}
