+++ sources/Colonization/AngelScriptBinders/Activities/HostActivity/BindHostActivityConstants.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindHostActivityConstants (Urho3D::Script *script);
+}
