+++ sources/Colonization/AngelScriptBinders/Activities/HostActivity/BindHostActivityEvents.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindHostActivityEvents (Urho3D::Script *script);
+}
