+++ sources/Colonization/Backend/MessagesHandler/MessagesHandler.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <Urho3D/Scene/Component.h>
 #include <Colonization/Backend/Player/Player.hpp>
-#include <Colonization/Activities/HostActivity.hpp>
+#include <Colonization/Activities/HostActivity/GameStateType.hpp>
 
 namespace Colonization
 {
