+++ sources/Colonization/Backend/MessagesHandler/NetworkMessageType.hpp
@@ -1,8 +1,4 @@
 #pragma once
-#include <Urho3D/Scene/Component.h>
-#include <Colonization/Backend/Player/Player.hpp>
-#include <Colonization/Activities/HostActivity.hpp>
-
 namespace Colonization
 {
 enum NetworkMessageType
